Migrate from dojo/store to dstore
This commit starts the migration from Dojo 1.x's dojo/store to Sitepen's
dstore which will provide the data management API for Dojo 2.x.

Models and property maps are introduced to allow offline data validation
and reduce the coupling with the REST API. Irrelevant properties sent by
the API are filtered. Inconsitently and missleadingly named properties
are mapped to proper names.